Refactor AlpacaBrokerage
- remove AlpacaApiBase class
- make AlpacaBrokerage partial class
- move DataQueueHandler implementation and private utility methods to partial files
