+++ Brokerages/Alpaca/AlpacaBrokerage.Utility.cs
@@ -14,249 +14,25 @@
 */
 
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using NodaTime;
 using QuantConnect.Brokerages.Alpaca.Markets;
-using QuantConnect.Data;
 using QuantConnect.Data.Market;
-using QuantConnect.Interfaces;
 using QuantConnect.Logging;
 using QuantConnect.Orders;
-using QuantConnect.Packets;
-using QuantConnect.Securities;
-using QuantConnect.Util;
 
 namespace QuantConnect.Brokerages.Alpaca
 {
     /// <summary>
-    /// Alpaca API base class
+    /// Alpaca Brokerage utility methods
     /// </summary>
-    public class AlpacaApiBase : Brokerage, IDataQueueHandler
+    public partial class AlpacaBrokerage
     {
-        private bool _isConnected;
-        private Thread _connectionMonitorThread;
-        private volatile bool _connectionLost;
-        private CancellationTokenSource _cancellationTokenSource = new CancellationTokenSource();
-
-        // Rest API requests must be limited to a maximum of 200 messages/minute
-        private readonly RateGate _messagingRateLimiter = new RateGate(200, TimeSpan.FromMinutes(1));
-
-        private RestClient _restClient;
-        private SockClient _sockClient;
-        private NatsClient _natsClient;
-
-        /// <summary>
-        /// This lock is used to sync 'PlaceOrder' and callback 'OnTransactionDataReceived'
-        /// </summary>
-        private readonly object _locker = new object();
-
-        /// <summary>
-        /// This container is used to keep pending to be filled market orders, so when the callback comes in we send the filled event
-        /// </summary>
-        protected readonly ConcurrentDictionary<int, Orders.OrderStatus> PendingFilledMarketOrders = new ConcurrentDictionary<int, Orders.OrderStatus>();
-
-        /// <summary>
-        /// The UTC time of the last received heartbeat message
-        /// </summary>
-        protected DateTime LastHeartbeatUtcTime;
-
-        /// <summary>
-        /// A lock object used to synchronize access to LastHeartbeatUtcTime
-        /// </summary>
-        protected readonly object LockerConnectionMonitor = new object();
-
-        /// <summary>
-        /// The list of ticks received
-        /// </summary>
-        protected readonly List<Tick> Ticks = new List<Tick>();
-
-        /// <summary>
-        /// The order provider
-        /// </summary>
-        protected IOrderProvider OrderProvider;
-
-        /// <summary>
-        /// The security provider
-        /// </summary>
-        protected ISecurityProvider SecurityProvider;
-
-        /// <summary>
-        /// The Alpaca api key
-        /// </summary>
-        protected string AccountKeyId;
-
-        /// <summary>
-        /// The Alpaca api secret
-        /// </summary>
-        protected string SecretKey;
-
-        /// <summary>
-        /// The Alpaca base url
-        /// </summary>
-        protected string BaseUrl;
-
-        /// <summary>
-        /// The market hours database
-        /// </summary>
-        protected MarketHoursDatabase MarketHours;
-
-        private readonly Dictionary<Symbol, DateTimeZone> _symbolExchangeTimeZones = new Dictionary<Symbol, DateTimeZone>();
-
-        public AlpacaApiBase(string name):base(name)
-        {
-
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AlpacaApiBase"/> class.
-        /// </summary>
-        /// <param name="orderProvider">The order provider.</param>
-        /// <param name="securityProvider">The holdings provider.</param>
-        /// <param name="keyId">The Alpaca api key id</param>
-        /// <param name="secretKey">The api secret key</param>
-        /// <param name="baseUrl">The Alpaca base url</param>
-        public void Initialize(IOrderProvider orderProvider, ISecurityProvider securityProvider, string keyId, string secretKey, string baseUrl)
-        {
-            OrderProvider = orderProvider;
-            SecurityProvider = securityProvider;
-            AccountKeyId = keyId;
-            SecretKey = secretKey;
-            BaseUrl = baseUrl;
-
-            MarketHours = MarketHoursDatabase.FromDataFolder();
-
-            // api client for alpaca
-            _restClient = new RestClient(AccountKeyId, SecretKey, baseUrl);
-
-            // websocket client for alpaca
-            _sockClient = new SockClient(AccountKeyId, SecretKey, BaseUrl);
-            _sockClient.OnTradeUpdate += OnTradeUpdate;
-            _sockClient.OnError += OnSockClientError;
-            _sockClient.ConnectAsync().SynchronouslyAwaitTask();
-
-            // polygon client for alpaca
-            _natsClient = new NatsClient(AccountKeyId, BaseUrl.Contains("staging"));
-            _natsClient.QuoteReceived += OnQuoteReceived;
-            _natsClient.TradeReceived += OnTradeReceived;
-            _natsClient.OnError += OnNatsClientError;
-            _natsClient.Open();
-        }
-
-        /// <summary>
-        /// Returns true if we're currently connected to the broker
-        /// </summary>
-        public override bool IsConnected => _isConnected && !_connectionLost;
-
-        /// <summary>
-        /// Connects the client to the broker's remote servers
-        /// </summary>
-        public override void Connect()
-        {
-            _isConnected = true;
-
-            // create new thread to manage disconnections and reconnections
-            _cancellationTokenSource = new CancellationTokenSource();
-            _connectionMonitorThread = new Thread(() =>
-            {
-                var nextReconnectionAttemptUtcTime = DateTime.UtcNow;
-                double nextReconnectionAttemptSeconds = 1;
-
-                lock (LockerConnectionMonitor)
-                {
-                    LastHeartbeatUtcTime = DateTime.UtcNow;
-                }
-
-                try
-                {
-                    while (!_cancellationTokenSource.IsCancellationRequested)
-                    {
-                        TimeSpan elapsed;
-                        lock (LockerConnectionMonitor)
-                        {
-                            elapsed = DateTime.UtcNow - LastHeartbeatUtcTime;
-                        }
-
-                        if (!_connectionLost && elapsed > TimeSpan.FromSeconds(20))
-                        {
-                            _connectionLost = true;
-                            nextReconnectionAttemptUtcTime = DateTime.UtcNow.AddSeconds(nextReconnectionAttemptSeconds);
-
-                            OnMessage(BrokerageMessageEvent.Disconnected("Connection with Alpaca server lost. " +
-                                                                         "This could be because of internet connectivity issues. "));
-                        }
-                        else if (_connectionLost)
-                        {
-                            try
-                            {
-                                if (elapsed <= TimeSpan.FromSeconds(20))
-                                {
-                                    _connectionLost = false;
-                                    nextReconnectionAttemptSeconds = 1;
-
-                                    OnMessage(BrokerageMessageEvent.Reconnected("Connection with Alpaca server restored."));
-                                }
-                                else
-                                {
-                                    if (DateTime.UtcNow > nextReconnectionAttemptUtcTime)
-                                    {
-                                        try
-                                        {
-                                            // check if we have a connection
-                                            GetInstrumentList();
-                                        }
-                                        catch (Exception)
-                                        {
-                                            // double the interval between attempts (capped to 1 minute)
-                                            nextReconnectionAttemptSeconds = Math.Min(nextReconnectionAttemptSeconds * 2, 60);
-                                            nextReconnectionAttemptUtcTime = DateTime.UtcNow.AddSeconds(nextReconnectionAttemptSeconds);
-                                        }
-                                    }
-                                }
-                            }
-                            catch (Exception exception)
-                            {
-                                Log.Error(exception);
-                            }
-                        }
-
-                        Thread.Sleep(1000);
-                    }
-                }
-                catch (Exception exception)
-                {
-                    Log.Error(exception);
-                }
-            })
-            { IsBackground = true };
-            _connectionMonitorThread.Start();
-            while (!_connectionMonitorThread.IsAlive)
-            {
-                Thread.Sleep(1);
-            }
-        }
-
-        /// <summary>
-        /// Disconnects the client from the broker's remote servers
-        /// </summary>
-        public override void Disconnect()
-        {
-            // request and wait for thread to stop
-            _cancellationTokenSource.Cancel();
-            _connectionMonitorThread?.Join();
-
-            _sockClient.DisconnectAsync().SynchronouslyAwaitTask();
-            _natsClient.Close();
-
-            _isConnected = false;
-        }
-
         /// <summary>
         /// Gets the list of available tradable instruments/products from Alpaca
         /// </summary>
-        public List<string> GetInstrumentList()
+        private List<string> GetInstrumentList()
         {
             try
             {
@@ -281,7 +57,7 @@ namespace QuantConnect.Brokerages.Alpaca
         /// </summary>
         /// <param name="instruments">the list of instruments to check</param>
         /// <returns>Dictionary containing the current quotes for each instrument</returns>
-        public Dictionary<string, Tick> GetRates(List<string> instruments)
+        private Dictionary<string, Tick> GetRates(List<string> instruments)
         {
             try
             {
@@ -302,126 +78,6 @@ namespace QuantConnect.Brokerages.Alpaca
                 throw;
             }
 
-        }
-
-        /// <summary>
-        /// Gets all open orders on the account.
-        /// NOTE: The order objects returned do not have QC order IDs.
-        /// </summary>
-        /// <returns>The open orders returned from Alpaca</returns>
-        public override List<Order> GetOpenOrders()
-        {
-            try
-            {
-                CheckRateLimiting();
-                var task = _restClient.ListOrdersAsync();
-                var orders = task.SynchronouslyAwaitTaskResult();
-
-                var qcOrders = new List<Order>();
-                foreach (var order in orders)
-                {
-                    qcOrders.Add(ConvertOrder(order));
-                }
-                return qcOrders;
-            }
-            catch (Exception e)
-            {
-                Log.Error(e);
-                if (e.InnerException != null)
-                    OnMessage(new BrokerageMessageEvent(BrokerageMessageType.Warning, 100, e.InnerException.Message));
-                throw;
-            }
-
-        }
-
-        /// <summary>
-        /// Gets all holdings for the account
-        /// </summary>
-        /// <returns>The current holdings from the account</returns>
-        public override List<Holding> GetAccountHoldings()
-        {
-            try
-            {
-                CheckRateLimiting();
-                var task = _restClient.ListPositionsAsync();
-                var holdings = task.SynchronouslyAwaitTaskResult();
-
-                var qcHoldings = new List<Holding>();
-                foreach (var holds in holdings)
-                {
-                    qcHoldings.Add(ConvertHolding(holds));
-                }
-
-                return qcHoldings;
-            }
-            catch (Exception e)
-            {
-                Log.Error(e);
-                if (e.InnerException != null)
-                    OnMessage(new BrokerageMessageEvent(BrokerageMessageType.Warning, 100, e.InnerException.Message));
-                throw;
-            }
-
-        }
-
-        /// <summary>
-        /// Gets the current cash balance for each currency held in the brokerage account
-        /// </summary>
-        /// <returns>The current cash balance for each currency available for trading</returns>
-        public override List<Cash> GetCashBalance()
-        {
-            try
-            {
-                CheckRateLimiting();
-                var task = _restClient.GetAccountAsync();
-                var balance = task.SynchronouslyAwaitTaskResult();
-
-                return new List<Cash>
-                    {
-                        new Cash("USD",
-                            balance.TradableCash,
-                            1m)
-                    };
-            }
-            catch (Exception e)
-            {
-                Log.Error(e);
-                if (e.InnerException != null)
-                    OnMessage(new BrokerageMessageEvent(BrokerageMessageType.Warning, 100, e.InnerException.Message));
-                throw;
-            }
-
-        }
-
-        /// <summary>
-        /// Places a new order and assigns a new broker ID to the order
-        /// </summary>
-        /// <param name="order">The order to be placed</param>
-        /// <returns>True if the request for a new order has been placed, false otherwise</returns>
-        public override bool PlaceOrder(Order order)
-        {
-            const int orderFee = 0;
-            order.PriceCurrency = "USD";
-
-            lock (_locker)
-            {
-                try
-                {
-                    var apOrder = GenerateAndPlaceOrder(order);
-                    order.BrokerId.Add(apOrder.OrderId.ToString());
-                }
-                catch (Exception e)
-                {
-                    Log.Error(e);
-                    if (e.InnerException != null)
-                        OnMessage(new BrokerageMessageEvent(BrokerageMessageType.Warning, 100, e.InnerException.Message));
-                    return false;
-                }
-
-            }
-            OnOrderEvent(new OrderEvent(order, DateTime.UtcNow, orderFee) { Status = Orders.OrderStatus.Submitted });
-
-            return true;
         }
 
         private IOrder GenerateAndPlaceOrder(Order order)
@@ -464,6 +120,7 @@ namespace QuantConnect.Brokerages.Alpaca
                 default:
                     throw new NotSupportedException("The order type " + order.Type + " is not supported.");
             }
+
             CheckRateLimiting();
             var task = _restClient.PostOrderAsync(order.Symbol.Value, quantity, side, type, timeInForce,
                 limitPrice, stopPrice);
@@ -474,58 +131,24 @@ namespace QuantConnect.Brokerages.Alpaca
         }
 
         /// <summary>
-        /// Updates the order with the same id
-        /// </summary>
-        /// <param name="order">The new order information</param>
-        /// <returns>True if the request was made for the order to be updated, false otherwise</returns>
-        public override bool UpdateOrder(Order order)
-        {
-            return false;
-        }
-
-        /// <summary>
-        /// Cancels the order with the specified ID
-        /// </summary>
-        /// <param name="order">The order to cancel</param>
-        /// <returns>True if the request was made for the order to be canceled, false otherwise</returns>
-        public override bool CancelOrder(Order order)
-        {
-            Log.Trace("AlpacaBrokerage.CancelOrder(): " + order);
-
-            if (!order.BrokerId.Any())
-            {
-                Log.Trace("AlpacaBrokerage.CancelOrder(): Unable to cancel order without BrokerId.");
-                return false;
-            }
-
-            foreach (var orderId in order.BrokerId)
-            {
-                CheckRateLimiting();
-                var task = _restClient.DeleteOrderAsync(new Guid(orderId));
-                task.SynchronouslyAwaitTaskResult();
-            }
-
-            return true;
-        }
-
-        /// <summary>
         /// Event handler for streaming events
         /// </summary>
         /// <param name="trade">The event object</param>
         private void OnTradeUpdate(ITradeUpdate trade)
         {
             Log.Trace("OnTransactionData: {0} {1} {2}", trade.Event, trade.Order.OrderId, trade.Order.OrderStatus);
+
             Order order;
             string tradeEvent = trade.Event.ToUpper();
             lock (_locker)
             {
-                order = OrderProvider.GetOrderByBrokerageId(trade.Order.OrderId.ToString());
+                order = _orderProvider.GetOrderByBrokerageId(trade.Order.OrderId.ToString());
             }
             if (order != null)
             {
                 if (tradeEvent == "FILL" || tradeEvent == "PARTIAL_FILL")
                 {
-                    order.PriceCurrency = SecurityProvider.GetSecurity(order.Symbol).SymbolProperties.QuoteCurrency;
+                    order.PriceCurrency = _securityProvider.GetSecurity(order.Symbol).SymbolProperties.QuoteCurrency;
 
                     var status = Orders.OrderStatus.Filled;
                     if (trade.Order.FilledQuantity < trade.Order.Quantity) status = Orders.OrderStatus.PartiallyFilled;
@@ -551,65 +174,6 @@ namespace QuantConnect.Brokerages.Alpaca
             }
         }
 
-        /// <summary>
-        /// Event handler for streaming quote ticks
-        /// </summary>
-        /// <param name="quote">The data object containing the received tick</param>
-        private void OnQuoteReceived(IStreamQuote quote)
-        {
-            LastHeartbeatUtcTime = DateTime.UtcNow;
-            var symbol = Symbol.Create(quote.Symbol, SecurityType.Equity, Market.USA);
-            var time = quote.Time;
-
-            // live ticks timestamps must be in exchange time zone
-            DateTimeZone exchangeTimeZone;
-            if (!_symbolExchangeTimeZones.TryGetValue(key: symbol, value: out exchangeTimeZone))
-            {
-                exchangeTimeZone = MarketHours.GetExchangeHours(Market.USA, symbol, SecurityType.Equity).TimeZone;
-                _symbolExchangeTimeZones.Add(symbol, exchangeTimeZone);
-            }
-            time = time.ConvertFromUtc(exchangeTimeZone);
-
-            var bidPrice = quote.BidPrice;
-            var askPrice = quote.AskPrice;
-            var tick = new Tick(time, symbol, bidPrice, bidPrice, askPrice);
-            tick.TickType = TickType.Quote;
-            tick.BidSize = quote.BidSize;
-            tick.AskSize = quote.AskSize;
-            lock (Ticks)
-            {
-                Ticks.Add(tick);
-            }
-        }
-
-        /// <summary>
-        /// Event handler for streaming trade ticks
-        /// </summary>
-        /// <param name="trade">The data object containing the received tick</param>
-        private void OnTradeReceived(IStreamTrade trade)
-        {
-            LastHeartbeatUtcTime = DateTime.UtcNow;
-            var symbol = Symbol.Create(trade.Symbol, SecurityType.Equity, Market.USA);
-            var time = trade.Time;
-
-            // live ticks timestamps must be in exchange time zone
-            DateTimeZone exchangeTimeZone;
-            if (!_symbolExchangeTimeZones.TryGetValue(key: symbol, value: out exchangeTimeZone))
-            {
-                exchangeTimeZone = MarketHours.GetExchangeHours(Market.USA, symbol, SecurityType.Equity).TimeZone;
-                _symbolExchangeTimeZones.Add(symbol, exchangeTimeZone);
-            }
-            time = time.ConvertFromUtc(exchangeTimeZone);
-
-            var tick = new Tick(time, symbol, trade.Price, trade.Price, trade.Price);
-            tick.TickType = TickType.Trade;
-            tick.Quantity = trade.Size;
-            lock (Ticks)
-            {
-                Ticks.Add(tick);
-            }
-        }
-
         private void OnNatsClientError(string error)
         {
             Log.Error($"NatsClient error: {error}");
@@ -629,7 +193,7 @@ namespace QuantConnect.Brokerages.Alpaca
         /// <param name="resolution">The requested resolution</param>
         /// <param name="requestedTimeZone">The requested timezone for the data</param>
         /// <returns>The list of bars</returns>
-        public IEnumerable<TradeBar> DownloadTradeBars(Symbol symbol, DateTime startTimeUtc, DateTime endTimeUtc, Resolution resolution, DateTimeZone requestedTimeZone)
+        private IEnumerable<TradeBar> DownloadTradeBars(Symbol symbol, DateTime startTimeUtc, DateTime endTimeUtc, Resolution resolution, DateTimeZone requestedTimeZone)
         {
             // This is due to the polygon API logic.
             // If start and end date is equal, then the result is null
@@ -738,7 +302,7 @@ namespace QuantConnect.Brokerages.Alpaca
         /// <param name="resolution">The requested resolution</param>
         /// <param name="requestedTimeZone">The requested timezone for the data</param>
         /// <returns>The list of bars</returns>
-        public IEnumerable<QuoteBar> DownloadQuoteBars(Symbol symbol, DateTime startTimeUtc, DateTime endTimeUtc, Resolution resolution, DateTimeZone requestedTimeZone)
+        private IEnumerable<QuoteBar> DownloadQuoteBars(Symbol symbol, DateTime startTimeUtc, DateTime endTimeUtc, Resolution resolution, DateTimeZone requestedTimeZone)
         {
             var period = resolution.ToTimeSpan();
 
@@ -854,7 +418,7 @@ namespace QuantConnect.Brokerages.Alpaca
         /// <param name="endTimeUtc">The ending time (UTC)</param>
         /// <param name="requestedTimeZone">The requested timezone for the data</param>
         /// <returns>The list of ticks</returns>
-        public IEnumerable<Tick> DownloadTicks(Symbol symbol, DateTime startTimeUtc, DateTime endTimeUtc, DateTimeZone requestedTimeZone)
+        private IEnumerable<Tick> DownloadTicks(Symbol symbol, DateTime startTimeUtc, DateTime endTimeUtc, DateTimeZone requestedTimeZone)
         {
             DateTime startTime = startTimeUtc;
             DateTime startTimeWithTZ = startTimeUtc.ConvertFromUtc(requestedTimeZone);
@@ -917,98 +481,6 @@ namespace QuantConnect.Brokerages.Alpaca
         }
 
         /// <summary>
-        /// Get the next ticks from the live trading data queue
-        /// </summary>
-        /// <returns>IEnumerable list of ticks since the last update.</returns>
-        public IEnumerable<BaseData> GetNextTicks()
-        {
-            lock (Ticks)
-            {
-                var copy = Ticks.ToArray();
-                Ticks.Clear();
-                return copy;
-            }
-        }
-
-        /// <summary>
-        /// Adds the specified symbols to the subscription
-        /// </summary>
-        /// <param name="job">Job we're subscribing for:</param>
-        /// <param name="symbols">The symbols to be added keyed by SecurityType</param>
-        public void Subscribe(LiveNodePacket job, IEnumerable<Symbol> symbols)
-        {
-            foreach (var symbol in symbols.Where(CanSubscribe))
-            {
-                Log.Trace($"AlpacaBrokerage.Subscribe(): {symbol}");
-
-                //CheckRateLimiting();
-                _natsClient.SubscribeQuote(symbol.Value);
-
-                //CheckRateLimiting();
-                _natsClient.SubscribeTrade(symbol.Value);
-            }
-        }
-
-        /// <summary>
-        /// Removes the specified symbols from the subscription
-        /// </summary>
-        /// <param name="job">Job we're processing.</param>
-        /// <param name="symbols">The symbols to be removed keyed by SecurityType</param>
-        public void Unsubscribe(LiveNodePacket job, IEnumerable<Symbol> symbols)
-        {
-            foreach (var symbol in symbols.Where(CanSubscribe))
-            {
-                Log.Trace($"AlpacaBrokerage.Unsubscribe(): {symbol}");
-
-                //CheckRateLimiting();
-                _natsClient.UnsubscribeQuote(symbol.Value);
-
-                //CheckRateLimiting();
-                _natsClient.UnsubscribeTrade(symbol.Value);
-            }
-        }
-
-        /// <summary>
-        /// Returns true if this brokerage supports the specified symbol
-        /// </summary>
-        private static bool CanSubscribe(Symbol symbol)
-        {
-            // ignore unsupported security types
-            if (symbol.ID.SecurityType != SecurityType.Equity)
-                return false;
-
-            if (symbol.Value.ToLower().IndexOf("universe") != -1)
-                return false;
-
-            return true;
-        }
-
-        /// <summary>
-        /// Subscribes to the requested symbols (using a single streaming session)
-        /// </summary>
-        /// <param name="symbolsToSubscribe">The list of symbols to subscribe</param>
-        protected void SubscribeSymbols(List<Symbol> symbolsToSubscribe)
-        {
-            var instruments = symbolsToSubscribe
-                .Select(symbol => symbol.Value)
-                .ToList();
-
-            foreach (var instrument in instruments)
-            {
-                _natsClient.SubscribeQuote(instrument);
-                _natsClient.SubscribeTrade(instrument);
-            }
-
-            //StopPricingStream();
-
-            //if (instruments.Count > 0)
-            //{
-            //    StartPricingStream(instruments);
-            //}
-        }
-
-
-        /// <summary>
         /// Converts an Alpaca order into a LEAN order.
         /// </summary>
         private Order ConvertOrder(IOrder order)
@@ -1058,7 +530,7 @@ namespace QuantConnect.Brokerages.Alpaca
             qcOrder.Status = Orders.OrderStatus.None;
             qcOrder.BrokerId.Add(id);
 
-            var orderByBrokerageId = OrderProvider.GetOrderByBrokerageId(id);
+            var orderByBrokerageId = _orderProvider.GetOrderByBrokerageId(id);
             if (orderByBrokerageId != null)
             {
                 qcOrder.Id = orderByBrokerageId.Id;
@@ -1075,7 +547,7 @@ namespace QuantConnect.Brokerages.Alpaca
         /// <summary>
         /// Converts an Alpaca position into a LEAN holding.
         /// </summary>
-        private Holding ConvertHolding(Markets.IPosition position)
+        private Holding ConvertHolding(IPosition position)
         {
             var securityType = SecurityType.Equity;
             var symbol = Symbol.Create(position.Symbol, securityType, Market.USA);
@@ -1090,22 +562,6 @@ namespace QuantConnect.Brokerages.Alpaca
                 Quantity = position.Quantity
             };
         }
-
-        /// <summary>
-        /// Returns the websocket client for alpaca
-        /// </summary>
-        //private SockClient GetSockClient()
-        //{
-        //    return new SockClient(AccountKeyId, SecretKey, BaseUrl);
-        //}
-
-        /// <summary>
-        /// Returns the polygon client for alpaca
-        /// </summary>
-        //private NatsClient GetNatsClient()
-        //{
-        //    return new NatsClient(AccountKeyId, BaseUrl.Contains("staging"));
-        //}
 
         private void CheckRateLimiting()
         {
